+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Improvements
 - Support range requests when downloading DICOMweb files ([#1444](../../pull/1444))
-- Bypass some scaling code when compositing multi sources ([#1447](../../pull/1447))
+- Bypass some scaling code when compositing multi sources ([#1447](../../pull/1447), [#1449](../../pull/1449))
 
 ## 1.27.1