+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Bypass some scaling code when compositing multi sources ([#1447](../../pull/1447), [#1449](../../pull/1449))
 - Do not create needless alpha bands in the multi source ([#1451](../../pull/1451))
 - Infer DICOM file size, when possible ([#1448](../../pull/1448))
+- Swap styles faster in the frame viewer ([#1452](../../pull/1452))
 
 ## 1.27.1